Tiny optomization
The teleport action was a lot of code for something simple. A function was also made for a single line of code...
The teleport function (now one instead of two and spread out in the teleport action) is now stored in the level!
The portals list and the list for portal pairs (bluePortal, pinkPortal, etc.) are now also stored in the level.